config: enable KEYS
This option provides support for retaining authentication tokens and
access keys in the kernel.

It also includes provision of methods by which such keys might be
associated with a process so that network filesystems, encryption
support and the like can find them.

Furthermore, a special type of key is available that acts as keyring:
a searchable sequence of keys. Each process is equipped with access
to five standard keyrings: UID-specific, GID-specific, session,
process and thread.